+++ plugins/wiggle/src/MultiLinearWiggleDisplay/models/model.tsx
@@ -34,6 +34,7 @@ import {
 } from '../../util'
 
 import Tooltip from '../components/Tooltip'
+import SharedWiggleMixin from '../../shared/modelShared'
 
 const randomColor = () =>
   '#000000'.replaceAll('0', () => (~~(Math.random() * 16)).toString(16))
@@ -68,85 +69,20 @@ export function stateModelFactory(
   return types
     .compose(
       'MultiLinearWiggleDisplay',
-      BaseLinearDisplay,
+      SharedWiggleMixin(configSchema),
       types.model({
         /**
          * #property
          */
         type: types.literal('MultiLinearWiggleDisplay'),
-        /**
-         * #property
-         */
-        configuration: ConfigurationReference(configSchema),
-        /**
-         * #property
-         */
-        selectedRendering: types.optional(types.string, ''),
-        /**
-         * #property
-         */
-        resolution: types.optional(types.number, 1),
-        /**
-         * #property
-         */
-        fill: types.maybe(types.boolean),
-        /**
-         * #property
-         */
-        minSize: types.maybe(types.number),
-        /**
-         * #property
-         */
-        color: types.maybe(types.string),
-        /**
-         * #property
-         */
-        posColor: types.maybe(types.string),
-        /**
-         * #property
-         */
-        negColor: types.maybe(types.string),
-        /**
-         * #property
-         */
-        summaryScoreMode: types.maybe(types.string),
-        /**
-         * #property
-         */
-        rendererTypeNameState: types.maybe(types.string),
-        /**
-         * #property
-         */
-        scale: types.maybe(types.string),
-        /**
-         * #property
-         */
-        autoscale: types.maybe(types.string),
-        /**
-         * #property
-         */
-        displayCrossHatches: types.maybe(types.boolean),
+
         /**
          * #property
          */
         layout: types.optional(types.frozen<Source[]>(), []),
-        /**
-         * #property
-         */
-        constraints: types.optional(
-          types.model({
-            max: types.maybe(types.number),
-            min: types.maybe(types.number),
-          }),
-          {},
-        ),
       }),
     )
     .volatile(() => ({
-      message: undefined as undefined | string,
-      stats: undefined as { scoreMin: number; scoreMax: number } | undefined,
-      statsRegion: undefined as string | undefined,
-      statsFetchInProgress: undefined as undefined | AbortController,
       featureUnderMouseVolatile: undefined as Feature | undefined,
       sourcesVolatile: undefined as Source[] | undefined,
     }))
@@ -163,21 +99,7 @@ export function stateModelFactory(
       clearLayout() {
         self.layout = []
       },
-      /**
-       * #action
-       */
-      updateQuantitativeStats(stats: { scoreMin: number; scoreMax: number }) {
-        const { scoreMin, scoreMax } = stats
-        const EPSILON = 0.000001
-        if (!self.stats) {
-          self.stats = { scoreMin, scoreMax }
-        } else if (
-          Math.abs(self.stats.scoreMax - scoreMax) > EPSILON ||
-          Math.abs(self.stats.scoreMin - scoreMin) > EPSILON
-        ) {
-          self.stats = { scoreMin, scoreMax }
-        }
-      },
+
       /**
        * #action
        */
@@ -186,121 +108,12 @@ export function stateModelFactory(
           self.sourcesVolatile = sources
         }
       },
-      /**
-       * #action
-       */
-      setColor(color: string) {
-        self.color = color
-      },
-      /**
-       * #action
-       */
-      setPosColor(color: string) {
-        self.posColor = color
-      },
-      /**
-       * #action
-       */
-      setNegColor(color: string) {
-        self.negColor = color
-      },
-      /**
-       * #action
-       */
-      setLoading(aborter: AbortController) {
-        const { statsFetchInProgress: statsFetch } = self
-        if (statsFetch !== undefined && !statsFetch.signal.aborted) {
-          statsFetch.abort()
-        }
-        self.statsFetchInProgress = aborter
-      },
-      /**
-       * #action
-       * this overrides the BaseLinearDisplayModel to avoid popping up a
-       * feature detail display, but still sets the feature selection on the
-       * model so listeners can detect a click
-       */
-      selectFeature(feature: Feature) {
-        const session = getSession(self)
-        if (isSelectionContainer(session)) {
-          session.setSelection(feature)
-        }
-      },
 
       /**
        * #action
        */
       setFeatureUnderMouse(f?: Feature) {
         self.featureUnderMouseVolatile = f
-      },
-      /**
-       * #action
-       */
-      setResolution(res: number) {
-        self.resolution = res
-      },
-      /**
-       * #action
-       */
-      setFill(fill: number) {
-        if (fill === 0) {
-          self.fill = true
-          self.minSize = 0
-        } else if (fill === 1) {
-          self.fill = false
-          self.minSize = 1
-        } else if (fill === 2) {
-          self.fill = false
-          self.minSize = 2
-        }
-      },
-      /**
-       * #action
-       */
-      toggleLogScale() {
-        self.scale = self.scale === 'log' ? 'linear' : 'log'
-      },
-      /**
-       * #action
-       */
-      setScaleType(scale?: string) {
-        self.scale = scale
-      },
-      /**
-       * #action
-       */
-      setSummaryScoreMode(val: string) {
-        self.summaryScoreMode = val
-      },
-
-      setAutoscale(val: string) {
-        self.autoscale = val
-      },
-
-      setMaxScore(val?: number) {
-        self.constraints.max = val
-      },
-
-      setRendererType(val: string) {
-        self.rendererTypeNameState = val
-      },
-      /**
-       * #action
-       */
-      setMinScore(val?: number) {
-        self.constraints.min = val
-      },
-      /**
-       * #action
-       */
-      toggleCrossHatches() {
-        self.displayCrossHatches = !self.displayCrossHatches
-      },
-      /**
-       * #action
-       */
-      setCrossHatches(cross: boolean) {
-        self.displayCrossHatches = cross
       },
     }))
     .views(self => ({
@@ -316,92 +129,17 @@ export function stateModelFactory(
       get TooltipComponent() {
         return Tooltip as unknown as React.FC
       },
-      /**
-       * #getter
-       */
-      get adapterTypeName() {
-        return self.adapterConfig.type
-      },
-      /**
-       * #getter
-       */
-      get rendererTypeNameSimple() {
-        return self.rendererTypeNameState || getConf(self, 'defaultRendering')
-      },
+
       /**
        * #getter
        */
       get rendererTypeName() {
-        const name = this.rendererTypeNameSimple
+        const name = self.rendererTypeNameSimple
         const rendererType = rendererTypes.get(name)
         if (!rendererType) {
           throw new Error(`unknown renderer ${name}`)
         }
         return rendererType
-      },
-
-      /**
-       * #getter
-       * subclasses can define these, as snpcoverage track does
-       */
-      get filters() {
-        return undefined
-      },
-
-      /**
-       * #getter
-       */
-      get scaleType() {
-        return self.scale ?? (getConf(self, 'scaleType') as string)
-      },
-      /**
-       * #getter
-       */
-      get maxScore() {
-        return self.constraints.max ?? (getConf(self, 'maxScore') as number)
-      },
-      /**
-       * #getter
-       */
-      get minScore() {
-        return self.constraints.min ?? (getConf(self, 'minScore') as number)
-      },
-    }))
-    .views(self => ({
-      /**
-       * #getter
-       */
-      get rendererConfig() {
-        const configBlob =
-          getConf(self, ['renderers', self.rendererTypeName]) || {}
-
-        const {
-          color,
-          displayCrossHatches,
-          fill,
-          minSize,
-          negColor,
-          posColor,
-          summaryScoreMode,
-          scaleType,
-        } = self
-
-        return self.rendererType.configSchema.create(
-          {
-            ...configBlob,
-            ...(scaleType ? { scaleType } : {}),
-            ...(fill !== undefined ? { filled: fill } : {}),
-            ...(displayCrossHatches !== undefined
-              ? { displayCrossHatches }
-              : {}),
-            ...(summaryScoreMode !== undefined ? { summaryScoreMode } : {}),
-            ...(color !== undefined ? { color } : {}),
-            ...(negColor !== undefined ? { negColor } : {}),
-            ...(posColor !== undefined ? { posColor } : {}),
-            ...(minSize !== undefined ? { minSize } : {}),
-          },
-          getEnv(self),
-        )
       },
     }))
     .views(self => ({
@@ -490,106 +228,32 @@ export function stateModelFactory(
               (!this.isMultiRow ? colors[i] || randomColor() : 'blue'),
           }))
       },
-      /**
-       * #getter
-       */
-      get filled(): boolean {
-        const { fill, rendererConfig } = self
-        return fill ?? readConfObject(rendererConfig, 'filled')
-      },
-      /**
-       * #getter
-       */
-      get summaryScoreModeSetting(): string {
-        const { summaryScoreMode: scoreMode, rendererConfig } = self
-        return scoreMode ?? readConfObject(rendererConfig, 'summaryScoreMode')
-      },
     }))
 
-    .views(self => {
-      let oldDomain: [number, number] = [0, 0]
-      return {
-        /**
-         * #getter
-         */
-        get domain() {
-          const { stats, scaleType, minScore, maxScore } = self
-          if (!stats) {
-            return undefined
-          }
-          const { scoreMin, scoreMax } = stats
+    .views(self => ({
+      /**
+       * #getter
+       */
+      get rowHeight() {
+        const { sources, height, isMultiRow } = self
+        return isMultiRow ? height / (sources?.length || 1) : height
+      },
+      /**
+       * #getter
+       */
+      get rowHeightTooSmallForScalebar() {
+        return this.rowHeight < 70
+      },
 
-          const ret = getNiceDomain({
-            domain: [scoreMin, scoreMax],
-            bounds: [minScore, maxScore],
-            scaleType,
-          })
-
-          // avoid weird scalebar if log value and empty region displayed
-          if (scaleType === 'log' && ret[1] === Number.MIN_VALUE) {
-            return [0, Number.MIN_VALUE]
-          }
-
-          // avoid returning a new object if it matches the old value
-          if (!deepEqual(oldDomain, ret)) {
-            oldDomain = ret
-          }
-
-          return oldDomain
-        },
-        /**
-         * #getter
-         */
-        get scaleOpts() {
-          const { scaleType, stats } = self
-          return {
-            autoscaleType: this.autoscaleType,
-            domain: this.domain,
-            stats,
-            scaleType,
-            inverted: getConf(self, 'inverted'),
-          }
-        },
-        /**
-         * #getter
-         */
-        get autoscaleType() {
-          return self.autoscale ?? (getConf(self, 'autoscale') as string)
-        },
-        /**
-         * #getter
-         */
-        get displayCrossHatchesSetting() {
-          const { displayCrossHatches, rendererConfig } = self
-          return (
-            displayCrossHatches ??
-            (readConfObject(rendererConfig, 'displayCrossHatches') as boolean)
-          )
-        },
-        /**
-         * #getter
-         */
-        get rowHeight() {
-          const { sources, height, isMultiRow } = self
-          return isMultiRow ? height / (sources?.length || 1) : height
-        },
-        /**
-         * #getter
-         */
-        get rowHeightTooSmallForScalebar() {
-          return this.rowHeight < 70
-        },
-
-        /**
-         * #getter
-         */
-        get useMinimalTicks() {
-          return (
-            getConf(self, 'minimalTicks') || this.rowHeightTooSmallForScalebar
-          )
-        },
-      }
-    })
+      /**
+       * #getter
+       */
+      get useMinimalTicks() {
+        return (
+          getConf(self, 'minimalTicks') || this.rowHeightTooSmallForScalebar
+        )
+      },
+    }))
     .views(self => ({
       /**
        * #getter
@@ -631,14 +295,6 @@ export function stateModelFactory(
           'darkred',
           'pink',
         ]
-      },
-
-      /**
-       * #getter
-       */
-      get adapterCapabilities() {
-        const { adapterTypeName } = self
-        return pluginManager.getAdapterType(adapterTypeName).adapterCapabilities
       },
     }))
     .views(self => {
@@ -720,70 +376,7 @@ export function stateModelFactory(
             ...superTrackMenuItems(),
             {
               label: 'Score',
-              subMenu: [
-                ...(self.hasResolution
-                  ? [
-                      {
-                        label: 'Resolution',
-                        subMenu: [
-                          {
-                            label: 'Finer resolution',
-                            onClick: () =>
-                              self.setResolution(self.resolution * 5),
-                          },
-                          {
-                            label: 'Coarser resolution',
-                            onClick: () =>
-                              self.setResolution(self.resolution / 5),
-                          },
-                        ],
-                      },
-                      {
-                        label: 'Summary score mode',
-                        subMenu: ['min', 'max', 'avg', 'whiskers'].map(elt => ({
-                          label: elt,
-                          type: 'radio',
-                          checked: self.summaryScoreModeSetting === elt,
-                          onClick: () => self.setSummaryScoreMode(elt),
-                        })),
-                      },
-                    ]
-                  : []),
-                {
-                  label:
-                    self.scaleType === 'log'
-                      ? 'Set linear scale'
-                      : 'Set log scale',
-                  onClick: () => self.toggleLogScale(),
-                },
-                {
-                  label: 'Autoscale type',
-                  subMenu: [
-                    ['local', 'Local'],
-                    ...(self.hasGlobalStats
-                      ? [
-                          ['global', 'Global'],
-                          ['globalsd', 'Global ± 3σ'],
-                        ]
-                      : []),
-                    ['localsd', 'Local ± 3σ'],
-                  ].map(([val, label]) => ({
-                    label,
-                    type: 'radio',
-                    checked: self.autoscaleType === val,
-                    onClick: () => self.setAutoscale(val),
-                  })),
-                },
-                {
-                  label: 'Set min/max score',
-                  onClick: () => {
-                    getSession(self).queueDialog(handleClose => [
-                      SetMinMaxDlg,
-                      { model: self, handleClose },
-                    ])
-                  },
-                },
-              ],
+              subMenu: self.scoreTrackMenuItems(),
             },
 
             ...(self.canHaveFill
