+++ plugins/wiggle/src/shared/modelShared.ts
@@ -1,72 +1,99 @@
-import { getConf } from '@jbrowse/core/configuration'
-import { Feature, getSession, isSelectionContainer } from '@jbrowse/core/util'
+import {
+  AnyConfigurationSchemaType,
+  ConfigurationReference,
+  getConf,
+  readConfObject,
+} from '@jbrowse/core/configuration'
+import {
+  Feature,
+  getEnv,
+  getSession,
+  isSelectionContainer,
+} from '@jbrowse/core/util'
+import { BaseLinearDisplay } from '@jbrowse/plugin-linear-genome-view'
 import { types } from 'mobx-state-tree'
+
+// locals
+import { getNiceDomain } from '../util'
+import { lazy } from 'react'
+
+// lazies
+const SetMinMaxDlg = lazy(() => import('./SetMinMaxDialog'))
 
 /**
  * #stateModel SharedWiggleModel
  */
-export default function SharedWiggleMixin() {
+export default function SharedWiggleMixin(
+  configSchema: AnyConfigurationSchemaType,
+) {
   return types
-    .model({
-      /**
-       * #property
-       */
-      selectedRendering: types.optional(types.string, ''),
-      /**
-       * #property
-       */
-      resolution: types.optional(types.number, 1),
-      /**
-       * #property
-       */
-      fill: types.maybe(types.boolean),
-      /**
-       * #property
-       */
-      minSize: types.maybe(types.number),
-      /**
-       * #property
-       */
-      color: types.maybe(types.string),
-      /**
-       * #property
-       */
-      posColor: types.maybe(types.string),
-      /**
-       * #property
-       */
-      negColor: types.maybe(types.string),
-      /**
-       * #property
-       */
-      summaryScoreMode: types.maybe(types.string),
-      /**
-       * #property
-       */
-      rendererTypeNameState: types.maybe(types.string),
-      /**
-       * #property
-       */
-      scale: types.maybe(types.string),
-      /**
-       * #property
-       */
-      autoscale: types.maybe(types.string),
-      /**
-       * #property
-       */
-      displayCrossHatches: types.maybe(types.boolean),
-      /**
-       * #property
-       */
-      constraints: types.optional(
-        types.model({
-          max: types.maybe(types.number),
-          min: types.maybe(types.number),
-        }),
-        {},
-      ),
-    })
+    .compose(
+      BaseLinearDisplay,
+      types.model({
+        /**
+         * #property
+         */
+        selectedRendering: types.optional(types.string, ''),
+        /**
+         * #property
+         */
+        resolution: types.optional(types.number, 1),
+        /**
+         * #property
+         */
+        fill: types.maybe(types.boolean),
+        /**
+         * #property
+         */
+        minSize: types.maybe(types.number),
+        /**
+         * #property
+         */
+        color: types.maybe(types.string),
+        /**
+         * #property
+         */
+        posColor: types.maybe(types.string),
+        /**
+         * #property
+         */
+        negColor: types.maybe(types.string),
+        /**
+         * #property
+         */
+        summaryScoreMode: types.maybe(types.string),
+        /**
+         * #property
+         */
+        rendererTypeNameState: types.maybe(types.string),
+        /**
+         * #property
+         */
+        scale: types.maybe(types.string),
+        /**
+         * #property
+         */
+        autoscale: types.maybe(types.string),
+        /**
+         * #property
+         */
+        displayCrossHatches: types.maybe(types.boolean),
+        /**
+         * #property
+         */
+        constraints: types.optional(
+          types.model({
+            max: types.maybe(types.number),
+            min: types.maybe(types.number),
+          }),
+          {},
+        ),
+        /**
+         * #property
+         */
+        configuration: ConfigurationReference(configSchema),
+      }),
+    )
     .volatile(() => ({
       message: undefined as undefined | string,
       stats: undefined as { scoreMin: number; scoreMax: number } | undefined,
@@ -228,6 +255,13 @@ export default function SharedWiggleMixin() {
 
       /**
        * #getter
+       */
+      get rendererTypeNameSimple() {
+        return self.rendererTypeNameState ?? getConf(self, 'defaultRendering')
+      },
+
+      /**
+       * #getter
        * subclasses can define these, as snpcoverage track does
        */
       get filters() {
@@ -253,6 +287,209 @@ export default function SharedWiggleMixin() {
        */
       get minScore() {
         return self.constraints.min ?? getConf(self, 'minScore')
+      },
+    }))
+    .views(self => ({
+      /**
+       * #getter
+       */
+      get adapterCapabilities() {
+        const type = self.adapterTypeName
+        const { pluginManager } = getEnv(self)
+        return pluginManager.getAdapterType(type).adapterCapabilities
+      },
+      /**
+       * #getter
+       */
+      get rendererConfig() {
+        const {
+          color,
+          displayCrossHatches,
+          fill,
+          minSize,
+          negColor,
+          posColor,
+          summaryScoreMode,
+          scaleType,
+          rendererTypeName,
+        } = self
+        const configBlob = getConf(self, ['renderers', rendererTypeName]) || {}
+        return self.rendererType.configSchema.create(
+          {
+            ...configBlob,
+            ...(scaleType ? { scaleType } : {}),
+            ...(fill !== undefined ? { filled: fill } : {}),
+            ...(displayCrossHatches !== undefined
+              ? { displayCrossHatches }
+              : {}),
+            ...(summaryScoreMode !== undefined ? { summaryScoreMode } : {}),
+            ...(color !== undefined ? { color } : {}),
+            ...(negColor !== undefined ? { negColor } : {}),
+            ...(posColor !== undefined ? { posColor } : {}),
+            ...(minSize !== undefined ? { minSize } : {}),
+          },
+          getEnv(self),
+        )
+      },
+
+      /**
+       * #getter
+       */
+      get autoscaleType() {
+        return self.autoscale ?? getConf(self, 'autoscale')
+      },
+    }))
+    .views(self => {
+      let oldDomain: [number, number] = [0, 0]
+      return {
+        /**
+         * #getter
+         */
+        get domain() {
+          const { stats, scaleType, minScore, maxScore } = self
+          if (!stats) {
+            return undefined
+          }
+
+          const ret = getNiceDomain({
+            domain: [stats.scoreMin, stats.scoreMax],
+            bounds: [minScore, maxScore],
+            scaleType,
+          })
+
+          // avoid weird scalebar if log value and empty region displayed
+          if (scaleType === 'log' && ret[1] === Number.MIN_VALUE) {
+            return [0, Number.MIN_VALUE]
+          }
+
+          // avoid returning a new object if it matches the old value
+          if (JSON.stringify(oldDomain) !== JSON.stringify(ret)) {
+            oldDomain = ret
+          }
+
+          return oldDomain
+        },
+      }
+    })
+    .views(self => ({
+      /**
+       * #getter
+       */
+      get filled(): boolean {
+        const { fill, rendererConfig } = self
+        return fill ?? readConfObject(rendererConfig, 'filled')
+      },
+      /**
+       * #getter
+       */
+      get summaryScoreModeSetting(): string {
+        const { summaryScoreMode: mode, rendererConfig } = self
+        return mode ?? readConfObject(rendererConfig, 'summaryScoreMode')
+      },
+
+      /**
+       * #getter
+       */
+      get scaleOpts() {
+        return {
+          domain: self.domain,
+          stats: self.stats,
+          autoscaleType: self.autoscaleType,
+          scaleType: self.scaleType,
+          inverted: getConf(self, 'inverted'),
+        }
+      },
+
+      /**
+       * #getter
+       */
+      get canHaveFill() {
+        return self.rendererTypeName === 'XYPlotRenderer'
+      },
+
+      /**
+       * #getter
+       */
+      get displayCrossHatchesSetting(): boolean {
+        const { displayCrossHatches: hatches, rendererConfig } = self
+        return hatches ?? readConfObject(rendererConfig, 'displayCrossHatches')
+      },
+      /**
+       * #getter
+       */
+      get hasResolution() {
+        return self.adapterCapabilities.includes('hasResolution')
+      },
+
+      /**
+       * #getter
+       */
+      get hasGlobalStats() {
+        return self.adapterCapabilities.includes('hasGlobalStats')
+      },
+    }))
+    .views(self => ({
+      scoreTrackMenuItems() {
+        return [
+          ...(self.hasResolution
+            ? [
+                {
+                  label: 'Resolution',
+                  subMenu: [
+                    {
+                      label: 'Finer resolution',
+                      onClick: () => self.setResolution(self.resolution * 5),
+                    },
+                    {
+                      label: 'Coarser resolution',
+                      onClick: () => self.setResolution(self.resolution / 5),
+                    },
+                  ],
+                },
+                {
+                  label: 'Summary score mode',
+                  subMenu: ['min', 'max', 'avg', 'whiskers'].map(elt => ({
+                    label: elt,
+                    type: 'radio',
+                    checked: self.summaryScoreModeSetting === elt,
+                    onClick: () => self.setSummaryScoreMode(elt),
+                  })),
+                },
+              ]
+            : []),
+          {
+            label:
+              self.scaleType === 'log' ? 'Set linear scale' : 'Set log scale',
+            onClick: () => self.toggleLogScale(),
+          },
+          {
+            label: 'Autoscale type',
+            subMenu: [
+              ['local', 'Local'],
+              ...(self.hasGlobalStats
+                ? [
+                    ['global', 'Global'],
+                    ['globalsd', 'Global ± 3σ'],
+                  ]
+                : []),
+              ['localsd', 'Local ± 3σ'],
+            ].map(([val, label]) => ({
+              label,
+              type: 'radio',
+              checked: self.autoscaleType === val,
+              onClick: () => self.setAutoscale(val),
+            })),
+          },
+          {
+            label: 'Set min/max score',
+            onClick: () => {
+              getSession(self).queueDialog(handleClose => [
+                SetMinMaxDlg,
+                { model: self, handleClose },
+              ])
+            },
+          },
+        ]
       },
     }))
 }
